refactor: reuse Farming Icon
