+++ src/components/Icons/index.ts
@@ -38,3 +38,4 @@ export { default as Telegram } from './Telegram'
 export { default as Velas } from './Velas'
 export { default as VelasLogoFull } from './VelasLogoFull'
 export { default as EthW } from './EthW'
+export { default as FarmingIcon } from './FarmingIcon'
